feat(otel): organize folder structure to pktvisor otel receiver
