+++ agent/otel/pktvisorreceiver/e2e_test.go
@@ -1,4 +1,4 @@
-package otel
+package pktvisorreceiver
 
 import (
 	"context"
