+++ agent/otel/pktvisorreceiver/testdata/config.yaml
@@ -1,0 +1,31 @@
+receivers:
+  pktvisor_prometheus:
+  pktvisor_prometheus/all_settings:
+    endpoint: "localhost:1234"
+    collection_interval: 30s
+    metrics_path: /v2/metrics
+    params:
+      columns: "name,messages"
+      key: [ "foo","bar" ]
+    use_service_account: true
+    tls_enabled: true
+    tls_config:
+      ca_file: "path"
+      cert_file: "path"
+      key_file: "path"
+      insecure_skip_verify: true
+
+processors:
+  nop:
+
+exporters:
+  nop:
+
+service:
+  pipelines:
+    metrics:
+      receivers:
+        - pktvisor_prometheus
+        - pktvisor_prometheus/all_settings
+      processors: [ nop ]
+      exporters: [ nop ]
