--- agent/otel/testdata/config.yaml
@@ -1,28 +1,0 @@
-receivers:
-  pktvisor_prometheus:
-    endpoint: "localhost:10854"
-    collection_interval: 10s
-    metrics_path: /api/v1/policies/__all/metrics/bucket/1
-    use_service_account: true
-
-
-processors:
-  nop:
-
-exporters:
-  prometheus:
-    endpoint: "localhost:8787"
-    namespace: test-space
-    send_timestamps: true
-    metric_expiration: 60m
-
-
-service:
-  pipelines:
-    metrics:
-      receivers:
-        - pktvisor_prometheus
-      processors:
-        - nop
-      exporters:
-        - prometheus
